Swaps: Popular Tokens (#5990)
